fix(imports): Correct imports of services with no data object
We previously we importing a data
object for each service regardless
if they had a data object or not.

In this PR we check all the exports
of the services file. From there we
import the services if they end
"Service" and the data objects if
they end with "Data"